Default FD implementations of get_*_du() functions
Previously, a driver had to either implement all get_*_du()
functions its problem required which are used in the ADRE class for the
exact jacobian, or just finite-difference the whole jacobian (really,
really slow). This allows the driver to implement only the get_*()
functions and still use the ADRE "exact" jacobian, which is much faster.
Quick tests indicate it's actually about the same speed as implementing
the exact derivative functions