use sparse matrix for fd jacobian
Instead of allocating and assigning to an n_dof*n_dof dense matrix
ignoring the fact that most entries are zero, we now use a sparse matrix
for the fd jacobian to which entries are added only if they are greater
than a predefined threshold (that can be customised). This is mostly
useful for custom (at the moment, non advection-diffusion-reaction
problems) that want a slightly more efficient fd jacobian.